feat(product): 添加商品搜索功能并更新相关路由和配置
本次提交主要实现了商品搜索功能，包括：
1. 在 `product.proto` 中添加了 `SearchProducts` RPC 方法及相关请求和响应消息。
2. 更新了 `productservice` 的客户端和服务端代码以支持搜索功能。
3. 新增了 `search_products.go` 和 `search_products_test.go` 文件，用于实现和测试搜索功能。
4. 更新了 `router` 相关文件，将路由注册逻辑集中到 `register.go` 中，并删除了旧的路由文件。
5. 更新了 `docker-compose-infra.yaml`，添加了 RabbitMQ 服务配置。
6. 修复了部分代码中的拼写错误和依赖问题。

这些改动旨在提高系统的可维护性和扩展性，同时为用户提供更丰富的商品搜索功能。

+++ app/payment/biz/service/update_payment_status.go
@@ -47,7 +47,7 @@ func (s *UpdatePaymentStatusService) Run(req *payment.UpdatePaymentStatusRequest
 		// TODO: 发布支付成功事件
 		// 用 mq 通知 order 修改状态，通知 prodcut 修改库存，通知 notifacation 发送通知
 		event := &payment.PaymentMQEvent {
-			PaymentId: paymentRecord.ID,
+			PaymentId: int64(paymentRecord.ID),
 			OrderId: paymentRecord.OrderId,
 			UserId: paymentRecord.UserId,
 			Amount: paymentRecord.Amount,
